requirements.txt
eyed3
pytube
Pillow
argparse